jwks/requirements.txt
functions-framework==3.*
google-cloud-storage
jwcrypto
google-cloud-logging
google-cloud-core
google-crc32c
google-cloud-secret-manager